requirements.txt
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-admin==2.0.2
django-excel-response2==3.0.6
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
frozendict==2.4.2
html5lib==1.1
idna==3.7
isoweek==1.3.3
lxml==5.2.1
multitasking==0.0.11
numpy==1.26.4
pandas==2.2.2
peewee==3.17.3
platformdirs==4.2.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
screen==1.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
webencodings==0.5.1
xlwt==1.3.0
yfinance==0.2.38
